# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "HouseFinder API",
    "description": "Lista degli endpoints dell'applicazione HouseFinder e informazioni sul loro utilizzo."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "security": [
    {
      "jwt": []
    }
  ],
  "securityDefinitions": {
    "jwt": {
      "name": "x-access-token",
      "type": "apiKey",
      "in": "header"
    }
  },
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/utente/registrazione": {
      "post": {
        "summary": "Registrazione",
        "description": "Registra un nuovo utente",
        "operationId": "registrazione",
        "tags": [
          "utente"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "Credenziali da registrare",
            "required": true,
            "schema": {
              "$ref": "#/definitions/registrazione_credenziali"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "La registrazione è avvenuta con successo."
          },
          "400": {
            "description": "L'input non è formattato correttamente."
          },
          "409": {
            "description": "Indirizzo email già presente nel sistema."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/utente/{id}": {
      "patch": {
        "summary": "Modifica profilo",
        "description": "Modifica le informazioni dell'account",
        "operationId": "modifica_profilo",
        "tags": [
          "utente"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Credenziali da modificare",
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "required": true,
            "description": "Credenziali da modificare",
            "schema": {
              "$ref": "#/definitions/registrazione_credenziali"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Modifica del profilo riuscita."
          },
          "404": {
            "description": "Utente non trovato."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "delete": {
        "summary": "Elimina profilo",
        "description": "Elimina l'account",
        "operationId": "cancella_account",
        "tags": [
          "utente"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "Id utente",
            "schema": {
              "properties": {
                "userId": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancellazione del profilo riuscita."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/utente/recupero-password": {
      "put": {
        "summary": "Recupero password",
        "description": "Recupero password utente",
        "operationId": "recupero_password",
        "tags": [
          "utente"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "required": true,
            "description": "Email account e password nuova",
            "schema": {
              "$ref": "#/definitions/login_credenziali"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password recuperata con successo."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
    "/utente/annunci-pubblicati": {
      "get": {
        "summary": "Annunci pubblicati",
        "description": "Lista degli annunci pubblicati dall'utente",
        "operationId": "get_annunci_pubblicati",
        "tags": [
          "utente"
        ],
        "responses": {
          "200": {
            "description": "Elenco degli annunci pubblicati ottenuto correttamente.",
            "schema": {
              "properties": {
                "annunci": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/annuncio"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Utente non autorizzato."
          },
          "404": {
            "description": "Utente non trovato."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/utente/annunci-salvati": {
      "get": {
        "summary": "Annunci salvati",
        "description": "Lista degli annunci salvati dall'utente",
        "operationId": "get_annunci_salvati",
        "tags": [
          "utente"
        ],
        "responses": {
          "200": {
            "description": "Elenco degli annunci salvati ottenuto correttamente.",
            "schema": {
              "properties": {
                "annunci": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/annuncio"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Utente non autorizzato."
          },
          "404": {
            "description": "Utente non trovato."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/utente/ricerche-salvate": {
      "get": {
        "summary": "Ricerche Salvate",
        "description": "Lista delle ricerche salvate dall'utente",
        "operationId": "get_ricerche_salvate",
        "tags": [
          "utente"
        ],
        "responses": {
          "200": {
            "description": "Elenco delle ricerche salvate ottenuto correttamente.",
            "schema": {
              "properties": {
                "annunci": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/ricerca"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Utente non autorizzato."
          },
          "404": {
            "description": "Utente non trovato."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/utente/chat": {
      "get": {
        "summary": "Lista chat",
        "description": "Lista delle chat dell'utente",
        "operationId": "get_chat",
        "tags": [
          "utente"
        ],
        "responses": {
          "200": {
            "description": "Elenco delle chat ottenuto correttamente.",
            "schema": {
              "properties": {
                "annunci": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/chat"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Utente non autorizzato."
          },
          "404": {
            "description": "Utente non trovato."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
    "/chat/{id}": {
      "get": {
        "summary": "Apri chat",
        "description": "Apri una chat",
        "operationId": "get_chat",
        "tags": [
          "chat"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id della chat",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Chat ottenuta correttamente."
          },
          "404": {
            "description": "Chat non trovata."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "utente": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string"
        },
        "cognome": {
          "type": "string"
        },
        "data_nascita": {
          "$ref": "#/definitions/date"
        },
        "numero_tel": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "annuncio": {
      "type": "object",
      "properties": {
        "creatore": {
          "type": "string"
        },
        "foto": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "superficie_tot": {
          "type": "number"
        },
        "numero_bagni": {
          "type": "number"
        },
        "numero_locali": {
          "type": "number"
        },
        "locali": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/locale"
          }
        },
        "prezzo": {
          "type": "number"
        },
        "classe_energetica": {
          "type": "string"
        },
        "indirizzo": {
          "type": "string"
        },
        "arredato": {
          "type": "boolean"
        },
        "vetrina": {
          "properties": {
            "data_inizio": {
              "$ref": "#/definitions/date"
            },
            "data_fine": {
              "$ref": "#/definitions/date"
            }
          }
        }
      }
    },
    "locale": {
      "type": "object",
      "properties": {
        "nome_locale": {
          "type": "string"
        },
        "superficie": {
          "type": "number"
        },
        "descrizione": {
          "type": "string"
        },
        "arredamento": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "ricerca": {
      "type": "object",
      "properties": {
        "testo": {
          "type": "string"
        },
        "filtri": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/filtro"
          }
        }
      }
    },
    "filtro": {
      "type": "object",
      "properties": {
        "min": {
          "type": "number"
        },
        "max": {
          "type": "number"
        }
      }
    },
    "chat": {
      "type": "object",
      "properties": {
        "utente": {
          "type": "object",
          "$ref": "#/definitions/utente"
        },
        "locatore": {
          "type": "object",
          "$ref": "#/definitions/utente"
        },
        "messaggi": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/messaggio"
          }
        },
        "annuncio": {
          "type": "object",
          "$ref": "#/definitions/annuncio"
        }
      }
    },
    "messaggio": {
      "type": "object",
      "properties": {
        "messaggio": {
          "type": "string"
        },
        "data": {
          "$ref": "#/definitions/date"
        },
        "mittente": {
          "type": "string"
        }
      }
    },
    "registrazione_credenziali": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string"
        },
        "cognome": {
          "type": "string"
        },
        "data_nascita": {
          "$ref": "#/definitions/date"
        },
        "numero_tel": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "login_credenziali": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
    "date": {
      "type": "string"
    }
  }
}